+++ CHANGELOG.md
@@ -10,8 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add support to realm update.
 - Add `device_registration_limit` field to `CreateRealm`, `GetRealmReply`
   and `UpdateRealm`.
+- Add `datastream_maximum_storage_retention` field to `CreateRealm`,
+  `GetRealmReply` and `UpdateRealm`.
 - Add `GetDeviceRegistrationLimit[Reply]` to retrieve the maximum
   number of registered devices per realm.
+- Add `GetDatastreamMaximumStorageRetention[Reply]` to retrieve the maximum
+  retention value for data in a realm.
 
 ### Changed
 - Bump Elixir to 1.15.7.